pyproject.toml
[tool.poetry]
name = "void-events"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
flask = "^1.1.2"
disputils = "^0.1.3"
Pyrebase4 = "^4.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
